decided to punt on user/pass api_key generation for now, also moved client code to pkg
