+++ README.md
@@ -8,8 +8,6 @@ Future state?  `In` can be used to get the security posture of a given applicati
 
 * `defectdojo_url`: *Required.* The path of the hosted instance of defectdojo.
 
-* `username`: *Required.* The username to authenticate to defectdojo with.
-
 * `api_key`: *Required.* Generated API key (for `username`) to interact with defectdojo, [see here](https://defectdojo.readthedocs.io/en/latest/api-v2-docs.html).
 
 * `debug`: *Optional.* Enable debug logging.
